split unit tests into two files
